requirements.txt
python-dotenv>=1.0.1
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.7.0
